Move methods from dims.reample to dims.grid and util
