+++ nlmod/gis.py
@@ -4,8 +4,7 @@ import os
 import geopandas as gpd
 import numpy as np
 
-from .dims.grid import polygons_from_model_ds
-from .dims.resample import get_affine_mod_to_world
+from .dims.grid import polygons_from_model_ds, get_affine_mod_to_world
 from .dims.layers import calculate_thickness
 
 logger = logging.getLogger(__name__)
